+++ addons/account/static/src/scss/bills_upload.scss
@@ -1,5 +1,4 @@
 .o_drop_area {
-    display: none;
     width: 100%;
     height: 100%;
     position: absolute;
