+++ addons/account/static/src/xml/legacy_bills_upload_views.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<templates>
+    <t t-name="account.DocumentsHiddenUploadForm">
+        <div class="d-none o_account_document_upload">
+            <t t-call="HiddenInputFile">
+                <t t-set="multi_upload" t-value="true"/>
+                <t t-set="fileupload_id" t-value="widget.fileUploadID"/>
+                <t t-set="fileupload_action" t-translation="off">/web/binary/upload_attachment</t>
+                <input type="hidden" name="model" value=""/>
+                <input type="hidden" name="id" value="0"/>
+            </t>
+        </div>
+    </t>
+
+    <t t-extend="ListView.buttons" t-name="DocumentsListView.buttons">
+        <t t-jquery="button.o_list_button_add" t-operation="after">
+            <button type="button" class="btn btn-secondary o_button_upload_document">
+                Upload
+            </button>
+        </t>
+    </t>
+
+    <t t-extend="KanbanView.buttons" t-name="DocumentsKanbanView.buttons">
+        <t t-jquery="button" t-operation="after">
+            <button type="button" class="btn btn-secondary o_button_upload_document">
+                Upload
+            </button>
+        </t>
+    </t>
+
+    <t t-name="account.DocumentDropZone">
+        <div class="o_drop_area">
+            <i class="fa fa-upload fa-10x"></i>
+        </div>
+    </t>
+</templates>
