+++ addons/account/views/account_journal_dashboard_view.xml
@@ -18,6 +18,7 @@
                         <div t-attf-class="#{!selection_mode ? kanban_color(record.color.raw_value) : ''}">
                             <t t-value="JSON.parse(record.kanban_dashboard.raw_value)" t-set="dashboard"/>
                             <t t-value="record.type.raw_value" t-set="journal_type"/>
+                            <t t-call="AccountDropZoneUploader"/>
                             <t t-call="JournalTop"/>
                             <div t-att-class="'container o_kanban_card_content' + (dashboard.is_sample_data ? ' o_sample_data' : '')">
                                 <div class="row">
@@ -34,6 +35,17 @@
                                 <t t-call="JournalManage"/>
                             </div>
                         </div>
+                    </t>
+
+                    <t t-name="AccountDropZoneUploader" owl="1">
+                        <AccountDropZone
+                            visible="state.dropzoneVisible"
+                            hideZone="() => state.dropzoneVisible = false"/>
+                        <AccountFileUploader
+                            extraContext="{
+                                default_journal_id: record.id.raw_value,
+                                default_move_type: journal_type == 'sale' and 'out_invoice' or journal_type == 'purchase' and 'in_invoice' or 'entry'
+                            }"/>
                     </t>
 
                     <t t-name="JournalTop">
